+++ src/CONSTANTS.js
@@ -27,6 +27,7 @@ const CONSTANTS = {
   UTXO_SELECTION_MAX_FEE_VS_SINGLE_UTXO_FEE_FACTOR: 5,
   MAX_STANDARD_TX_SIZE: 100000,
   MAX_P2SH_SIGOPS: 15,
+  COINBASE_MATURITY: 100,
   // limit to how many times an unconfirmed input in a new tx can be respent
   UTXO_CHAINED_SPENDING_LIMIT_FOR_TX: 25,
   FEES: {
@@ -40,6 +41,10 @@ const CONSTANTS = {
   },
   UNCONFIRMED_TRANSACTION_STATUS_CODE: -1,
   WALLET_TYPES: {
+    ADDRESS: 'address',
+    PUBLICKEY: 'publicKey',
+    PRIVATEKEY: 'privateKey',
+    // TODO: DEPRECATE.
     SINGLE_ADDRESS: 'single_address',
     HDWALLET: 'hdwallet',
     HDPUBLIC: 'hdpublic',
@@ -70,11 +75,11 @@ const CONSTANTS = {
       'generateAddress',
       'getPrivateKeys',
       'injectPlugin',
-      'getIdentityHDKeyByIndex',
     ],
     UNSAFE_PROPERTIES: [
       'storage',
       'keyChain',
+      'identities',
     ],
     SAFE_PROPERTIES: [
       'offlineMode',
@@ -86,6 +91,13 @@ const CONSTANTS = {
       'strategy',
       'network',
     ],
+  },
+  TRANSACTION_HISTORY_TYPES: {
+    RECEIVED: 'received',
+    SENT: 'sent',
+    ADDRESS_TRANSFER: 'address_transfer',
+    ACCOUNT_TRANSFER: 'account_transfer',
+    UNKNOWN: 'unknown',
   },
   STORAGE: {
     autosaveIntervalTime: 10 * 1000,