+++ src/plugins/Worker.js
@@ -1,7 +1,6 @@
 const _ = require('lodash');
 const logger = require('../logger');
 const StandardPlugin = require('./StandardPlugin');
-const { WorkerFailedOnExecute, WorkerFailedOnStart } = require('../errors');
 
 // eslint-disable-next-line no-underscore-dangle
 const _defaultOpts = {
@@ -18,6 +17,10 @@ class Worker extends StandardPlugin {
     this.worker = null;
     this.workerPass = 0;
     this.isWorkerRunning = false;
+
+    this.awaitOnInjection = _.has(opts, 'awaitOnInjection')
+      ? opts.awaitOnInjection
+      : false;
 
     this.firstExecutionRequired = _.has(opts, 'firstExecutionRequired')
       ? opts.firstExecutionRequired
@@ -66,20 +69,35 @@ class Worker extends StandardPlugin {
 
       if (this.executeOnStart) await this.execWorker();
     } catch (e) {
-      throw new WorkerFailedOnStart(this.name, e);
+      this.emit('error', e, {
+        type: 'plugin',
+        pluginType: 'worker',
+        pluginName: this.name,
+      });
     }
   }
 
-  async stopWorker(reason = null) {
-    let payloadResult = reason;
+  /**
+   * @param {Object} [options]
+   * @param {Boolean} [options.force=false]
+   * @param {Boolean} [options.reason]
+   * @returns {Promise<void>}
+   */
+  async stopWorker(options = {}) {
+    let payloadResult = options.reason;
+
     clearInterval(this.worker);
+
     this.worker = null;
     this.workerPass = 0;
     this.isWorkerRunning = false;
+
     const eventType = `WORKER/${this.name.toUpperCase()}/STOPPED`;
+
     if (this.onStop) {
-      payloadResult = await this.onStop();
+      payloadResult = await this.onStop(options);
     }
+
     this.state.started = false;
     logger.debug(JSON.stringify({ eventType, result: payloadResult }));
     this.parentEvents.emit(eventType, { type: eventType, payload: payloadResult });
@@ -100,9 +118,15 @@ class Worker extends StandardPlugin {
       try {
         payloadResult = await this.execute();
       } catch (e) {
-        await this.stopWorker(e.message);
+        await this.stopWorker({
+          reason: e.message,
+        });
 
-        throw new WorkerFailedOnExecute(this.name, e);
+        this.emit('error', e, {
+          type: 'plugin',
+          pluginType: 'worker',
+          pluginName: this.name,
+        });
       }
     } else {
       throw new Error(`Worker ${this.name}: Missing execute function`);