+++ src/plugins/Workers/TransactionSyncStreamWorker/TransactionSyncStreamWorker.js
@@ -1,8 +1,15 @@
 const {
   Transaction, MerkleBlock, InstantLock,
 } = require('@dashevo/dashcore-lib');
+const GrpcError = require('@dashevo/grpc-common/lib/server/error/GrpcError');
+const GrpcErrorCodes = require('@dashevo/grpc-common/lib/server/error/GrpcErrorCodes');
+const { WALLET_TYPES } = require('../../../CONSTANTS');
+const sleep = require('../../../utils/sleep');
 
 const Worker = require('../../Worker');
+const isBrowser = require('../../../utils/isBrowser');
+
+const logger = require('../../../logger');
 
 class TransactionSyncStreamWorker extends Worker {
   constructor(options) {
@@ -10,6 +17,7 @@ class TransactionSyncStreamWorker extends Worker {
       name: 'TransactionSyncStreamWorker',
       executeOnStart: true,
       firstExecutionRequired: true,
+      awaitOnInjection: true,
       workerIntervalTime: 0,
       gapLimit: 10,
       dependencies: [
@@ -30,6 +38,8 @@ class TransactionSyncStreamWorker extends Worker {
     this.syncIncomingTransactions = false;
     this.stream = null;
     this.incomingSyncPromise = null;
+    this.pendingRequest = {};
+    this.delayedRequests = {};
   }
 
   /**
@@ -123,7 +133,15 @@ class TransactionSyncStreamWorker extends Worker {
     // instead of usual injection process
     const {
       skipSynchronizationBeforeHeight,
+      skipSynchronization,
     } = (this.storage.store.syncOptions || {});
+
+    if (skipSynchronization) {
+      logger.debug('TransactionSyncStreamWorker - Wallet created from a new mnemonic. Sync from the best block height.');
+      const bestBlockHeight = this.storage.store.chains[this.network.toString()].blockHeight;
+      this.setLastSyncedBlockHeight(bestBlockHeight);
+      return;
+    }
 
     if (skipSynchronizationBeforeHeight) {
       this.setLastSyncedBlockHeight(
@@ -150,20 +168,96 @@ class TransactionSyncStreamWorker extends Worker {
     this.incomingSyncPromise = this.startIncomingSync();
   }
 
-  async onStop() {
+  /**
+   * @param {Object} [options]
+   * @param {Boolean} [options.force=false]
+   * @param {Boolean} [options.reason]
+   *
+   * @returns {Promise<boolean>}
+   */
+  async onStop(options = {}) {
+    // in case of disconnect we don't need to wait until the wallet complete the sync process
+    if (options.force) {
+      this.pendingRequest = {};
+    }
+
+    // Sync, will require transaction and their blockHeader to be fetched before resolving.
+    // As await onStop() is a way to wait for execution before continuing,
+    // this ensure onStop will properly let the plugin to warn about all
+    // completion of pending request.
+    if (Object.keys(this.pendingRequest).length !== 0) {
+      await sleep(200);
+
+      return this.onStop();
+    }
+
     this.syncIncomingTransactions = false;
 
-    if (this.stream) {
-      this.stream.cancel();
-      // When calling stream.cancel(), the stream will emit 'error' event with the code 'CANCELLED'.
-      // There are two cases when this happens: when the gap limit is filled and syncToTheGapLimit
-      // and the stream needs to be restarted with new parameters, and here,
-      // when stopping the worker.
-      // The code in stream worker distinguishes whether it need to reconnect or not by the fact
-      // that the old stream object is present or not. When it is set to null, it won't try to
-      // reconnect to the stream.
-      this.stream = null;
+    if (isBrowser()) {
+      // Under browser environment, grpc-web doesn't call error and end events
+      // so we call it by ourselves
+      if (this.stream) {
+        return new Promise((resolve) => setImmediate(() => {
+          if (this.stream) {
+            this.stream.cancel();
+
+            const error = new GrpcError(GrpcErrorCodes.CANCELLED, 'Cancelled on client');
+            // call onError events
+            this.stream.f.forEach((func) => func(error));
+
+            // call onEnd events
+            this.stream.c.forEach((func) => func());
+
+            this.stream = null;
+          }
+
+          resolve(true);
+        }));
+      }
     }
+
+    // Wrapping `cancel` in `setImmediate` due to bug with double-free
+    // explained here (https://github.com/grpc/grpc-node/issues/1652)
+    // and here (https://github.com/nodejs/node/issues/38964)
+    return new Promise((resolve) => setImmediate(() => {
+      if (this.stream) {
+        this.stream.cancel();
+        // When calling stream.cancel(), the stream will emit 'error' event
+        // with the code 'CANCELLED'.
+        // There are two cases when this happens: when the gap limit is filled
+        // and syncToTheGapLimit and the stream needs to be restarted with new parameters,
+        // and here, when stopping the worker.
+        // The code in stream worker distinguishes whether it need to reconnect or not by the fact
+        // that the old stream object is present or not. When it is set to null, it won't try to
+        // reconnect to the stream.
+        this.stream = null;
+      }
+      resolve(true);
+    }));
+  }
+
+  setLastSyncedBlockHash(hash) {
+    const { walletId } = this;
+    const accountsStore = this.storage.store.wallets[walletId].accounts;
+
+    const accountStore = ([WALLET_TYPES.HDWALLET, WALLET_TYPES.HDPUBLIC].includes(this.walletType))
+      ? accountsStore[this.BIP44PATH.toString()]
+      : accountsStore[this.index.toString()];
+
+    accountStore.blockHash = hash;
+
+    return accountStore.blockHash;
+  }
+
+  getLastSyncedBlockHash() {
+    const { walletId } = this;
+    const accountsStore = this.storage.store.wallets[walletId].accounts;
+
+    const { blockHash } = ([WALLET_TYPES.HDWALLET, WALLET_TYPES.HDPUBLIC].includes(this.walletType))
+      ? accountsStore[this.BIP44PATH.toString()]
+      : accountsStore[this.index.toString()];
+
+    return blockHash;
   }
 }
 
@@ -174,6 +268,8 @@ TransactionSyncStreamWorker.prototype.setLastSyncedBlockHeight = require('./meth
 TransactionSyncStreamWorker.prototype.getLastSyncedBlockHash = require('./methods/getLastSyncedBlockHash');
 TransactionSyncStreamWorker.prototype.getLastSyncedBlockHeight = require('./methods/getLastSyncedBlockHeight');
 TransactionSyncStreamWorker.prototype.startHistoricalSync = require('./methods/startHistoricalSync');
+TransactionSyncStreamWorker.prototype.handleTransactionFromStream = require('./methods/handleTransactionFromStream');
+TransactionSyncStreamWorker.prototype.processChunks = require('./methods/processChunks');
 TransactionSyncStreamWorker.prototype.startIncomingSync = require('./methods/startIncomingSync');
 TransactionSyncStreamWorker.prototype.syncUpToTheGapLimit = require('./methods/syncUpToTheGapLimit');