+++ src/plugins/Workers/TransactionSyncStreamWorker/methods/syncUpToTheGapLimit.js
@@ -1,4 +1,8 @@
 const logger = require('../../../../logger');
+const onStreamEnd = require('../handlers/onStreamEnd');
+const onStreamError = require('../handlers/onStreamError');
+const onStreamData = require('../handlers/onStreamData');
+const Queue = require('../../../../utils/Queue/Queue');
 
 /**
  *
@@ -14,6 +18,7 @@ module.exports = async function syncUpToTheGapLimit({
 }) {
   const self = this;
   const addresses = this.getAddressesToSync();
+  self.addresses = addresses;
   logger.debug(`syncing up to the gap limit: - from block: ${fromBlockHash || fromBlockHeight} Count: ${count}`);
 
   if (fromBlockHash == null && fromBlockHeight == null) {
@@ -34,51 +39,22 @@ module.exports = async function syncUpToTheGapLimit({
     throw new Error('Limited to one stream at the same time.');
   }
   self.stream = stream;
-  let reachedGapLimit = false;
+  self.network = network;
+  self.hasReachedGapLimit = false;
+  // The order is important, however, some async calls are being performed
+  // in order to additionally fetch metadata for each valid tx chunks.
+  // We therefore need to temporarily store chunks for handling.
+  self.chunksQueue = new Queue();
 
-  return new Promise((resolve, reject) => {
+  // eslint-disable-next-line no-async-promise-executor
+  return new Promise(async (resolve, reject) => {
+    // handler for error being thrown for job processing
+    self.chunksQueue.on('error', (error) => {
+      reject(error);
+    });
     stream
-      .on('data', async (response) => {
-        /* First check if any instant locks appeared */
-        const instantLocksReceived = this.constructor.getInstantSendLocksFromResponse(response);
-        instantLocksReceived.forEach((isLock) => {
-          this.importInstantLock(isLock);
-        });
-
-        /* Incoming transactions handling */
-        const transactionsFromResponse = this.constructor
-          .getTransactionListFromStreamResponse(response);
-        const walletTransactions = this.constructor
-          .filterWalletTransactions(transactionsFromResponse, addresses, network);
-
-        if (walletTransactions.transactions.length) {
-          const addressesGeneratedCount = await self
-            .importTransactions(walletTransactions.transactions);
-
-          reachedGapLimit = reachedGapLimit || addressesGeneratedCount > 0;
-
-          if (reachedGapLimit) {
-            logger.silly('TransactionSyncStreamWorker - end stream - new addresses generated');
-            // If there are some new addresses being imported
-            // to the storage, that mean that we hit the gap limit
-            // and we need to update the bloom filter with new addresses,
-            // i.e. we need to open another stream with a bloom filter
-            // that contains new addresses.
-
-            // DO not setting null this.stream allow to know we
-            // need to reset our stream (as we pass along the error)
-            stream.cancel();
-          }
-        }
-      })
-      .on('error', (err) => {
-        logger.silly(`TransactionSyncStreamWorker - end stream on error: ${err}`);
-        reject(err);
-      })
-      .on('end', () => {
-        logger.silly('TransactionSyncStreamWorker - end stream on request');
-        self.stream = null;
-        resolve(reachedGapLimit);
-      });
+      .on('data', (data) => onStreamData(self, data))
+      .on('error', (error) => onStreamError(error, reject))
+      .on('end', () => onStreamEnd(self, resolve));
   });
 };