+++ src/types/Account/Account.d.ts
@@ -8,11 +8,12 @@ import {
     PrivateKey,
     Strategy,
     Network,
-    Plugins, RawTransaction, StatusInfo, TransactionsMap, WalletObj
+    broadcastTransactionOpts,
+    Plugins, RawTransaction, TransactionsMap, WalletObj, StatusInfo, TransactionsHistory
 } from "../types";
 import { KeyChain } from "../KeyChain/KeyChain";
-import { HDPrivateKey, InstantLock } from "@dashevo/dashcore-lib";
-import { Wallet } from "../../index";
+import { InstantLock } from "@dashevo/dashcore-lib";
+import { Identities, Wallet} from "../../index";
 import { Transport } from "../../transport/Transport";
 import { BlockHeader } from "@dashevo/dashcore-lib/typings/block/BlockHeader";
 import { UnspentOutput } from "@dashevo/dashcore-lib/typings/transaction/UnspentOutput";
@@ -35,33 +36,33 @@ export declare class Account {
     store: Storage.store;
     walletId: string;
     transport: Transport;
+    identities: Identities;
 
     isReady(): Promise<boolean>;
     isInitialized(): Promise<boolean>;
     getBIP44Path(network?: Network, index?: number): string;
     getNetwork(): Network;
 
-    broadcastTransaction(rawtx: Transaction|RawTransaction): Promise<transactionId>;
+    broadcastTransaction(rawtx: Transaction|RawTransaction, options?: broadcastTransactionOpts): Promise<transactionId>;
     connect(): boolean;
     createTransaction(opts: Account.createTransactionOptions): Transaction;
     decode(method: string, data: any): any;
-    decrypt(method: string, data: any, secret: string, encoding: "hex"|string): string;
+    decrypt(method: string, data: any, secret: string, encoding?: "hex"|string): string;
+    encrypt(method: string, data: any, secret: string): string;
     disconnect(): Promise<Boolean>;
-    fetchAddressInfo(addresObj: AddressObj, fetchUtxo: boolean): Promise<AddressInfo | false>;
-    fetchStatus(): Promise<StatusInfo|{blocks:number}>
+    fetchAddressInfo(addressObj: AddressObj, fetchUtxo: boolean): Promise<AddressInfo | false>;
+    fetchStatus(): Promise<StatusInfo>;
     forceRefreshAccount(): boolean;
     generateAddress(path: string): AddressObj;
     getAddress(index: number, _type: AddressType): AddressObj;
     getAddresses(_type: AddressType): [AddressObj];
     getBlockHeader(identifier: string|number):Promise<BlockHeader>
     getConfirmedBalance(displayDuffs?: boolean): number;
-    getIdentityHDKeyById(identityId: string, keyIndex: number): HDPrivateKey;
-    getIdentityHDKeyByIndex(identityIndex: number, keyIndex: number): HDPrivateKey;
-    getIdentityIds(): string[];
     getPlugin(name: string): Object;
     getPrivateKeys(addressList: [PublicAddress]): [PrivateKey];
     getTotalBalance(displayDuffs?: boolean): number;
     getTransaction(txid: transactionId): Transaction;
+    getTransactionHistory(): TransactionsHistory;
     getTransactions(): [Transaction];
     getUTXOS(): [UnspentOutput];
     getUnconfirmedBalance(displayDuffs?: boolean): number;
@@ -71,13 +72,16 @@ export declare class Account {
     hasPlugins([Plugin]): {found:Boolean, results:[{name: string}]};
     injectPlugin(unsafePlugin: Plugins, allowSensitiveOperation?: boolean, awaitOnInjection?: boolean): Promise<any>;
     sign(object: Transaction, privateKeys: [PrivateKey], sigType?: number): Transaction;
-    waitForInstantLock(string: transactionHash): Promise<InstantLock>;
+    waitForInstantLock(transactionHash: string): Promise<InstantLock>;
 }
 
 export declare interface RecipientOptions {
     satoshis?: number;
     amount?: number;
     address: PublicAddress,
+}
+export declare interface getUTXOSOptions {
+    coinbaseMaturity?: number;
 }
 
 export declare namespace Account {