+++ src/types/Wallet/Wallet.d.ts
@@ -1,8 +1,8 @@
-import {Mnemonic, PrivateKey, HDPublicKey, Strategy, Network, Plugins, AddressInfoMap, WalletType} from "../types";
-import {Account} from "../Account/Account";
-import {Storage} from "../Storage/Storage";
-import {HDPrivateKey} from "@dashevo/dashcore-lib";
-import {Transport} from "../../transport/Transport";
+import { Mnemonic, PrivateKey, PublicKey, PublicAddress, Address, HDPublicKey, Network, Plugins } from "../types";
+import { Account } from "../Account/Account";
+import { Storage } from "../Storage/Storage";
+import { HDPrivateKey } from "@dashevo/dashcore-lib";
+import { Transport } from "../../transport/Transport";
 
 export declare class Wallet {
     offlineMode: boolean;
@@ -59,7 +59,10 @@ export declare namespace Wallet {
         privateKey?: PrivateKey | string;
         HDPrivateKey?: HDPrivateKey | string;
         HDPublicKey?: HDPublicKey | string;
+        publicKey?: PublicKey | string;
+        address?: Address | PublicAddress | string;
         unsafeOptions?: IWalletUnsafeOptions;
+        waitForInstantLockTimeout?: number;
     }
 
     interface IWalletUnsafeOptions {